requirements.txt
undetected-chromedriver>=3.5.0
selenium>=4.15.0
selenium-stealth>=1.0.6
python-telegram-bot>=20.0
APScheduler>=3.10.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pytz>=2024.1
